+++ internal/elasticsearch/ml/calendar_event/models_test.go
@@ -190,7 +190,7 @@ func TestParseCompositeID(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			calID, evtID, diags := parseCompositeID(tt.id)
+			calID, evtID, diags := parseCalendarEventFullCompositeID(tt.id)
 			if tt.expectError {
 				assert.True(t, diags.HasError())
 			} else {